backend/api/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework>=3.13.1
django-cors-headers>=3.13.0
djangorestframework-simplejwt[crypto]>=5.2.0